requirements.txt
streamlit==1.28.1
google-generativeai==0.3.1
anthropic==0.7.1
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0
